Moved mailing functionality to a service, #1.
