+++ src/Lithnet.Laps.Web/Lithnet.Laps.Web/Mail/IMailer.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace Lithnet.Laps.Web.Mail
+{
+    public interface IMailer
+    {
+        void SendEmail(IEnumerable<string> recipients, string subject, string body);
+    }
+}
