+++ src/Lithnet.Laps.Web/Lithnet.Laps.Web/Web.config
@@ -21,6 +21,7 @@
       <register type="Lithnet.Laps.Web.Security.Authorization.ConfigurationFile.IAvailableTargets, Lithnet.Laps.Web" mapTo="Lithnet.Laps.Web.Security.Authorization.ConfigurationFile.AvailableTargets, Lithnet.Laps.Web" />
       <register type="Lithnet.Laps.Web.Audit.IReporting, Lithnet.Laps.Web" mapTo="Lithnet.Laps.Web.Audit.Reporting, Lithnet.Laps.Web" />
       <register type="Lithnet.Laps.Web.IRateLimiter, Lithnet.Laps.Web" mapTo="Lithnet.Laps.Web.RateLimiter, Lithnet.Laps.Web" />
+	  <register type="Lithnet.Laps.Web.Mail.IMailer, Lithnet.Laps.Web" mapTo="Lithnet.Laps.Web.Mail.SmtpMailer, Lithnet.Laps.Web" />
 
       <!-- The mapping for ILapsConfig is defined in Startup.cs -->
     </container>
